move HasRemoteAddr trait to reth-net-common
