+++ crates/net/common/src/lib.rs
@@ -9,3 +9,5 @@
 
 pub mod ban_list;
 pub mod bandwidth_meter;
+/// Traits related to tokio streams
+pub mod stream;
